moved pages to 'pages' folder
