+++ src/pages/PeoplePage/index.tsx
